pyproject.toml
[project]
name = "meeting-copilot"
version = "0.1.0"
description = "Live meeting copilot: transcript → short replies"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "websockets>=12.0",
  "sounddevice>=0.4.6",
  "numpy>=1.26",
  "langdetect>=1.0.9",
  "pyperclip>=1.8.2",
  "python-dotenv>=1.0.1",
  "pydantic>=2.7.0",
  "rich>=13.7.1",
  "anyio>=4.4.0"
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "ruff>=0.5.0", "mypy>=1.10.0"]

[tool.ruff]
line-length = 100
